+++ ebayscraper/src/constants.py
